// src/app/classes/personality.enum.ts
export enum PersonalityEnum {
  model_citizen = 'Model Citizen',
  model_professional = 'Model Professional',
  professional = 'Professional',
  perfectionist = 'Perfectionist	',
  temperamental = 'Temperamental	',
  driven = 'Driven',
  determined = 'Determined',
  slack = 'Slack',
  casual = 'Casual',
  very_ambitious = 'Very Ambitious',
  amibitous = 'Amibitous',
  unambitious = 'Unambitious',
  honest = 'Honest',
  sporting = 'Sporting',
  easily_discouraged = 'Easily Discouraged',
  low_determination = 'Low Determination	',
  unsporting = 'Unsporting',
  realist = 'Realist',
  very_loyal = 'Very Loyal',
  loyal = 'Loyal',
  iron_willed = 'Iron Willed',
  resilient = 'Resilient',
  spineless = 'Spineless',
  low_self_belief = 'Low Self-Belief',
  light_hearted = 'Light-Hearted',
  spirited = 'Spirited',
  jovial = 'Jovial',
  resolute = 'Resolute',
  fairly_professional = 'Fairly Professional',
  fairly_determined = 'Fairly Determined	',
  fairly_amibitous = 'Fairly Amibitous	',
  fairly_loyal = 'Fairly Loyal',
  fairly_sporting = 'Fairly Sporting	',
  balanced = 'Balanced',
}